完善跨进程信号控制

+++ lshcurator/workers/base.py
@@ -18,20 +18,27 @@ from queue import Empty, Queue
 from time import sleep
 from typing import Callable
 
+from ..config import WorkerConfig
 from ..utils.types import WorkerSlot, WorkerReport
 
 
 class WorkerBase(ABC):
     """Base class for workers."""
-    def __init__(self, *, worker_id: int, report_queue: multiprocessing.Queue[WorkerReport], **kwargs):
+    def __init__(self, *, worker_config: WorkerConfig, **kwargs):
         """
         Args:
             worker_id (int): Worker 的唯一标识符，主进程通过此 ID 来区分不同的 worker 进程。
             report_queue (multiprocessing.Queue): 用于向主进程发送报告的队列，worker 进程通过此队列将状态和结果报告给主进程。
         """
         super().__init__(**kwargs)
-        self.worker_id: int = worker_id
-        self.report_queue: multiprocessing.Queue[WorkerReport] = report_queue
+        self._worker_config: WorkerConfig = worker_config
+
+    @property
+    def worker_id(self) -> int: return self._worker_config.worker_id
+    @property
+    def report_queue(self) -> multiprocessing.Queue[WorkerReport]: return self._worker_config.report_queue
+    @property
+    def stop_event(self) -> multiprocessing.Event: return self._worker_config.stop_event
 
     @abstractmethod
     def job(self, **kwargs):
@@ -45,17 +52,16 @@ class WorkerBase(ABC):
 
 def _run_worker(
     worker_cls: Callable[..., WorkerBase],
-    worker_id: int,
-    report_queue: multiprocessing.Queue[WorkerReport],
+    worker_config: WorkerConfig,
     worker_init_kwargs: dict,
     job_kwargs: dict,
 ):
     """Worker 进程的入口函数，负责创建 Worker 实例并执行任务"""
     try:
-        worker = worker_cls(worker_id=worker_id, report_queue=report_queue, **worker_init_kwargs)  # 创建 Worker 实例，传递 worker_id 和 report_queue 以及其他必要参数
+        worker = worker_cls(worker_config=worker_config, **worker_init_kwargs)  # 创建 Worker 实例，传递 worker_id 和 report_queue 以及其他必要参数
         worker.job(**job_kwargs)  # 执行 Worker 的具体任务逻辑
     except Exception as e:
-        report_queue.put(WorkerReport(worker_id=worker_id, status='error', message=f'Worker encountered an error: {str(e)}'))  # 向主进程报告错误
+        worker_config.report_queue.put(WorkerReport(worker_id=worker_config.worker_id, status='error', message=f'Worker encountered an error: {str(e)}'))  # 向主进程报告错误
 
 
 class WorkerManagerBase(ABC):
@@ -113,17 +119,22 @@ class WorkerManagerBase(ABC):
     def add_subprocess(self, worker_cls: Callable[..., WorkerBase], worker_init_kwargs: dict | None = None, job_kwargs: dict | None = None) -> int:
         """创建一个新的 worker slot 并添加到队列中，target 是 worker 进程的入口函数，args 和 kwargs 是传递给 target 的参数"""
         worker_id = self._allocate_worker_slot_id()  # 自动分配一个新的 worker slot ID，确保不与现有 ID 冲突
+        worker_stop_event = self._multiprocessing_context.Event()  # 创建一个新的事件对象，用于通知 worker 进程停止
         self.add_worker_slot(WorkerSlot(
             process=self._multiprocessing_context.Process(
                 target=_run_worker,
                 kwargs={
                     'worker_cls': worker_cls,
-                    'worker_id': worker_id,
-                    'report_queue': self._worker_report_queue,  # 将报告队列传递给 worker 进程，确保 worker 可以通过此队列向主进程发送报告
+                    'worker_config': WorkerConfig(
+                        stop_event=worker_stop_event,
+                        report_queue=self._worker_report_queue,
+                        worker_id=worker_id,
+                    ),
                     'worker_init_kwargs': worker_init_kwargs or {},  # 传递给 Worker 实例的初始化参数
                     'job_kwargs': job_kwargs or {},  # 传递给 Worker.job 方法的参数
                 }
             ),
+            stop_event=worker_stop_event,
             worker_id=worker_id,
         ))
         return worker_id
@@ -138,7 +149,7 @@ class WorkerManagerBase(ABC):
             try: slot: WorkerSlot = self._add_subprocess_queue.get(timeout=1)  # 等待 1 秒获取添加请求，避免无限阻塞
             except Empty: continue  # 队列为空，继续等待
 
-            # 确认管理器仍然在运行
+            # 再次确认管理器仍然在运行
             if not self._add_subprocess_handler_is_running: break
 
             if slot.worker_id is None: slot.worker_id = self._allocate_worker_slot_id()  # 自动分配一个新的 idx，确保不与现有 idx 冲突
@@ -148,10 +159,9 @@ class WorkerManagerBase(ABC):
 
     def remove_worker_slot(self, worker_id: int) -> None:
         """移除指定 idx 的 worker slot"""
-        with self._worker_slots_lock:  # 防止同时释放同一 worker slot 导致的竞争条件
+        with self._worker_slots_lock:  # 全程持锁防止同时释放同一 worker slot 导致的竞争条件
             if worker_id not in self._worker_slots: return  # worker_id 不存在，无需移除
-            process = self._worker_slots[worker_id].process
-            self.stop_subprocesses(process=process)  # 停止 worker 进程，确保资源被正确释放
+            self.stop_subprocesses(worker_id=worker_id)  # 停止 worker 进程，确保资源被正确释放
             del self._worker_slots[worker_id]  # 从 worker_slots 中移除 worker slot
             self._worker_slots_ids.remove(worker_id)  # 从 worker slot ID 列表中移除 worker_id，确保未来分配的新 ID 不会与已存在的 ID 冲突
 
@@ -169,10 +179,17 @@ class WorkerManagerBase(ABC):
         """处理 worker 报告的具体逻辑"""
         raise NotImplementedError('Subclasses must implement work_report_handler to process worker reports')
 
-    @staticmethod
-    def stop_subprocesses(process: multiprocessing.Process):
-        """停止指定 worker 进程，确保资源被正确释放"""
-        if process.is_alive():
+    def stop_subprocesses(self, worker_id: int):
+        """
+        停止指定 worker 进程，确保资源被正确释放
+        注意：
+        全程不持锁，需要上游保证不会同时调用 stop_subprocesses 来停止同一个 worker_id，避免竞争条件导致的异常行为
+        """
+        slot = self._worker_slots.get(worker_id, None)
+        if slot is None: return  # worker_id 不存在，无需停止
+        slot.stop_event.set()  # 设置停止事件，通知 worker 进程尽快完成当前任务并退出
+        process = slot.process
+        if process.is_alive():  # 确保进程仍然存活后才尝试等待和终止，避免不必要的操作
             process.join(timeout=8)  # 等待 worker 进程结束
             if process.is_alive():
                 process.terminate()  # 强制终止仍然存活的 worker 进程
@@ -187,6 +204,7 @@ class WorkerManagerBase(ABC):
         """等待所有 worker 进程结束，并停止报告处理线程"""
         self._add_subprocess_handler_is_running = False  # 停止添加 worker slot 的线程
         self._add_subprocess_thread.join()  # 等待添加 worker slot 的线程结束
+        self._add_subprocess_queue.shutdown(immediate=True)  # 关闭添加 worker slot 的队列，确保没有未处理的添加请求
 
         for slot_idx in self.worker_slots.keys(): self.remove_worker_slot(slot_idx)  # 移除 worker slot 时会自动停止对应的 worker 进程
